requirements.txt
async-timeout==5.0.1
bidict==0.23.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.8
colorama==0.4.6
cryptography==3.4.8
DateTime==5.5
dnspython==2.1.0
et_xmlfile==2.0.0
eventlet==0.33.0
Flask==2.0.1
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
playwright==1.57.0
pycparser==2.22
pyee==13.0.0
PyMySQL==1.0.2
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.5.0
pytz==2025.1
RapidFuzz==3.14.3
redis==6.4.0
requests==2.28.1
simple-websocket==1.1.0
six==1.17.0
spotipy==2.22.1
SQLAlchemy==1.4.44
timedelta==2020.12.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.20
waitress==3.0.2
Werkzeug==2.0.1
wsproto==1.2.0
zope.event==5.1.1
zope.interface==7.2
